+++ sdk/resourcemover/arm-resourcemover/src/models/parameters.ts
@@ -99,7 +99,7 @@ export const moveCollectionName: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2023-08-01",
+    defaultValue: "2024-01-01",
     isConstant: true,
     serializedName: "api-version",
     type: {