+++ sdk/resourcemover/arm-resourcemover/src/resourceMoverServiceAPI.ts
@@ -74,7 +74,7 @@ export class ResourceMoverServiceAPI extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-resourcemover/2.2.0`;
+    const packageDetails = `azsdk-js-arm-resourcemover/2.3.0`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -127,7 +127,7 @@ export class ResourceMoverServiceAPI extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2023-08-01";
+    this.apiVersion = options.apiVersion || "2024-01-01";
     this.moveCollections = new MoveCollectionsImpl(this);
     this.moveResources = new MoveResourcesImpl(this);
     this.unresolvedDependencies = new UnresolvedDependenciesImpl(this);